move API response objects in sub-package of model
