+++ service/src/main/java/gov/nasa/pds/api/registry/model/RequestAndResponseContext.java
@@ -26,6 +26,10 @@ import gov.nasa.pds.api.registry.exceptions.ApplicationTypeException;
 import gov.nasa.pds.api.registry.exceptions.LidVidNotFoundException;
 import gov.nasa.pds.api.registry.exceptions.NothingFoundException;
 import gov.nasa.pds.api.registry.exceptions.UnknownGroupNameException;
+import gov.nasa.pds.api.registry.model.api_responses.Pds4ProductBusinessObject;
+import gov.nasa.pds.api.registry.model.api_responses.PdsProductBusinessObject;
+import gov.nasa.pds.api.registry.model.api_responses.ProductBusinessLogic;
+import gov.nasa.pds.api.registry.model.api_responses.WyriwygBusinessObject;
 import gov.nasa.pds.api.registry.model.identifiers.LidVidUtils;
 import gov.nasa.pds.api.registry.model.identifiers.PdsLidVid;
 import gov.nasa.pds.api.registry.model.identifiers.PdsProductIdentifier;
